# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Quickbuy API",
        "description": "API documentation for Quickbuy backend system.",
        "contact": {
            "email": "support@Quickbuy.com"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/profile/photo": {
            "put": {
                "tags": [
                    "Users"
                ],
                "summary": "Update user's profile photo",
                "operationId": "8dcd865ec5ea0f36ce3a26f9ca77baf8",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "photo"
                                ],
                                "properties": {
                                    "photo": {
                                        "description": "The new profile photo to upload",
                                        "type": "string",
                                        "format": "binary"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Profile photo updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Profile photo updated successfully"
                                        },
                                        "photo_url": {
                                            "type": "string",
                                            "example": "https://yourdomain.com/storage/profile_photos/1/avatar.png"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "errors": {
                                            "properties": {
                                                "photo": {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "string",
                                                        "example": "The photo must be an image."
                                                    }
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
...
}